backend/requirements.txt
Flask==3.0.3
flask_cors
python-dotenv==1.0.1
# google-generativeai==0.8.0
# openai==1.13.3
# langchain==0.1.11
# langchain-openai==0.0.8
langchain-google-genai==2.0.9
pillow == 11.0.0
